requirements.txt
flake8==5.0.4
mccabe==0.7.0
numpy==1.23.2
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
pytesseract==0.3.10
python-dotenv==0.20.0
yapf==0.32.0
